+++ example/lib/disable_form_sample.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class DisableFormSampleScreen extends StatefulWidget {
+  @override
+  _DisableFormSampleScreenState createState() =>
+      _DisableFormSampleScreenState();
+}
+
+class _DisableFormSampleScreenState extends State<DisableFormSampleScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}